rename package for consistency
