+++ apm-agent-plugins/apm-cmd-plugin/src/main/java/co/elastic/apm/agent/process/ProcessExitInstrumentation.java
@@ -22,7 +22,7 @@
  * under the License.
  * #L%
  */
-package co.elastic.apm.agent.cmd;
+package co.elastic.apm.agent.process;
 
 import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.method.MethodDescription;
