+++ apm-agent-plugins/apm-cmd-plugin/src/main/java/co/elastic/apm/agent/process/ProcessHelper.java
@@ -22,9 +22,8 @@
  * under the License.
  * #L%
  */
-package co.elastic.apm.agent.cmd;
+package co.elastic.apm.agent.process;
 
-import co.elastic.apm.agent.annotation.NonnullApi;
 import co.elastic.apm.agent.bci.VisibleForAdvice;
 import co.elastic.apm.agent.impl.transaction.Span;
 import co.elastic.apm.agent.impl.transaction.TraceContextHolder;
