+++ apm-agent-plugins/apm-cmd-plugin/src/main/java/co/elastic/apm/agent/process/ProcessStartInstrumentation.java
@@ -22,7 +22,7 @@
  * under the License.
  * #L%
  */
-package co.elastic.apm.agent.cmd;
+package co.elastic.apm.agent.process;
 
 import co.elastic.apm.agent.impl.transaction.TraceContextHolder;
 import net.bytebuddy.asm.Advice;
