+++ apm-agent-plugins/apm-cmd-plugin/src/main/resources/META-INF/services/co.elastic.apm.agent.bci.ElasticApmInstrumentation
@@ -1,2 +1,2 @@
-co.elastic.apm.agent.cmd.ProcessStartInstrumentation
-co.elastic.apm.agent.cmd.ProcessExitInstrumentation
+co.elastic.apm.agent.process.ProcessStartInstrumentation
+co.elastic.apm.agent.process.ProcessExitInstrumentation
